+++ apis/user.js
@@ -6,32 +6,50 @@ const userRoute = express.Router();
 
 
 /**
- *  @swagger 
- * /user/getUser:
- *   get:
- *     description: Get a list of all users
- *     produces:
- *       - application/json
- *     responses:
- *       200:
- *         description: Successfully retrieved list of users
- *       404:
- *         description: No users found
- *       500:
- *         description: Internal server error
- */
+*  @swagger 
+* /user:
+*   get:
+*     description: Get a list of all users
+*     tags: [User]
+*     produces:
+*       - application/json
+*     responses:
+*       200:
+*         description: Success
+*         content:
+*           application/json:
+*            schema:
+*              type: object
+*              properties:
+*                email:
+*                 type: string
+*                password:
+*                 type: string
+*       404:
+*         description: No data
+*         content:
+*           text/plain:
+*             schema:
+*               type: string
+*               example: 'No data found'
+*/
 
-userRoute.get("/getUser", async (req, res) => {
-    const data = await userService.getUser();
+userRoute.get("/", async (req, res) => {
+    let data = await userService.getUser();
+    if(data.length>0)
     res.send(data);
+    else{
+     res.status(404).json("No data found");   
+    }
 })
 
 
 /**
 * @swagger
-* /user/getUser/{id}:
+* /user/{id}:
 *   get:
 *     description: Retrieve user from the server.
+*     tags: [User]
 *     parameters:
 *       - name: id
 *         in: path
@@ -42,24 +60,59 @@ userRoute.get("/getUser", async (req, res) => {
 *          - application/json   
 *     responses:
 *       200:
-*         description: Successfully retrieved user
+*         description: Success
+*         content:
+*           application/json:
+*            schema:
+*              type: object
+*              properties:
+*                email:
+*                 type: string
+*                password:
+*                 type: string
 *       404:
-*         description: No user found
+*         description: No data
+*         content:
+*           text/plain:
+*             schema:
+*               type: string
+*               example: 'No data found'
 *       500:
 *         description: Internal server error
+*       400:
+*         description: Invalid Id
+*         content:
+*           text/plain:
+*             schema:
+*               type: string
+*               example: 'Invalid Id format'
 *    
 */
 
-userRoute.get("/getUser/:id", async (req, res) => {
-    const user = await userService.getUserById(req.params.id);
-    res.send(user);
-
+userRoute.get("/:id", async (req, res) => {
+    let isValidObjectId = (id) => {
+        const objectIdRegex = /^[0-9a-fA-F]{24}$/;
+      
+        return objectIdRegex.test(id);
+      };
+      
+      if (isValidObjectId(req.params.id)) {
+        let employee = await userService.getUserById(req.params.id);
+        if(employee.length>0)
+        res.send(employee);
+        else
+         res.status(404).json("No data found");
+    
+      } else {
+        res.status(400).json("Invalid Id Format");
+      }
 })
 
 /**
 * @swagger
-* /user/createUser:
+* /user:
 *   post:
+*     tags: [User]
 *     requestBody:
 *       required: true
 *       content:
@@ -80,17 +133,17 @@ userRoute.get("/getUser/:id", async (req, res) => {
 *         description: Bad request
 */
 
-userRoute.post("/createUser", async (req, res) => {
-    const data = await userService.createUser(req.body);
+userRoute.post("/", async (req, res) => {
+    let data = await userService.createUser(req.body);
     res.send(data);
-
 })
 
 /**
 * @swagger
-*  /user/editUser/{id}:
+*  /user/{id}:
 *  put:
 *     description: Update an existing user using its ID.
+*     tags: [User]
 *     parameters:
 *       - name: id
 *         in: path
@@ -117,35 +170,58 @@ userRoute.post("/createUser", async (req, res) => {
 *         description: Internal server error
 */
 
-userRoute.put("/editUser/:id", async (req, res) => {
-    const data = await userService.editUser(req.body, req.params.id);
+userRoute.put("/:id", async (req, res) => {
+    let data = await userService.editUser(req.body, req.params.id);
     res.send(data)
 })
 
 /**
 * @swagger
-* /user/removeUser/{id}:
+* /user/{id}:
 *  delete:
- *     description: delete an existing user using its ID.
+*     description: delete an existing user using its ID.
+*     tags: [User]
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         schema:
 *          type: string
- *     responses:
+*     responses:
 *       '200':
-*         description: user deleted successfully
+*         description: Success
+*         content:
+*           text/plain:
+*             schema:
+*               type: string
+*               example: 'user deleted successfully' 
 *       '404':
-*         description: user not found
+*         description: No data found
+*         content:
+*           text/plain:
+*             schema:
+*               type: string
+*               example: 'No data found' 
 *       '500':
 *         description: Internal server error     
 */
 
-userRoute.delete("/removeUser/:id", async (req, res) => {
-    await userService.deleteUser(req.params.id);
-    res.send("User deleted succcessfully");
+userRoute.delete("/:id", async (req, res) => {
+    let isValidObjectId = (id) => {
+        const objectIdRegex = /^[0-9a-fA-F]{24}$/;  
+    return objectIdRegex.test(id);
+};
 
+  if (isValidObjectId(req.params.id)) {
+  let data=await userService.deleteUser(req.params.id);
+  if(data.deletedCount>0)
+  res.send("user deleted successfully");
+  else
+  res.status(404).json("No data found");
+
+  } else {
+  res.status(400).json("Invalid Id Format");
+  }
 })
 
 module.exports = userRoute;
